requirements.txt
bcrypt
fastapi
loguru
passlib
pydantic
python-jose
uvicorn
tortoise-orm
requests
jsonpath
pytest